Move xhr mocking to main flow
